% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{basicColData.tiny}
\alias{basicColData.tiny}
\title{basicColData.tiny: metadata about a tiny collection of files for demonstrating TxRegInfra2}
\format{
DataFrame from S4Vectors
}
\usage{
basicColData.tiny
}
\description{
basicColData.tiny: metadata about a tiny collection of files for demonstrating TxRegInfra2
}
\examples{
data(basicColData.tiny)
head(basicColData.tiny)
}
\keyword{datasets}
