% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongolite.R
\name{prune_to_server_line}
\alias{prune_to_server_line}
\title{utility to prune away header lines on mongodb command line query}
\usage{
prune_to_server_line(x, tag = "MongoDB server")
}
\arguments{
\item{x}{character() vector}

\item{tag}{character(1) token to be found up to and including which we discard results from mongo shell}
}
\value{
character vector
}
\description{
utility to prune away header lines on mongodb command line query
}
