% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tnt_genetrack_hg19}
\alias{tnt_genetrack_hg19}
\title{gene addresses for visualization with TnTGenome}
\format{
TnT GeneTrack instance
}
\usage{
tnt_genetrack_hg19
}
\description{
gene addresses for visualization with TnTGenome
}
\note{
only includes genes with symbols in org.Hs.eg.db
}
\keyword{datasets}
