% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tnt_txtrack_hg19}
\alias{tnt_txtrack_hg19}
\title{transcript addresses for visualization with TnTGenome}
\format{
TnT TxTrack instance
}
\usage{
tnt_txtrack_hg19
}
\description{
transcript addresses for visualization with TnTGenome
}
\keyword{datasets}
